% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{add_matrix_plate_with_specimens}
\alias{add_matrix_plate_with_specimens}
\title{Add Matrix Plate with Specimens}
\usage{
add_matrix_plate_with_specimens(
  plate_uid,
  location_id,
  specimen_entries,
  create_missing_specimens = FALSE,
  create_missing_subjects = FALSE
)
}
\arguments{
\item{plate_uid}{unique identifier for matrix plate}

\item{location_id}{integer ID for plate location}

\item{specimen_entries}{list of specimen entries}

\item{create_missing_specimens}{Logical whether or not to create missing
specimens}

\item{create_missing_subjects}{Logical whether or not to create missing study
subjects}
}
\value{
returns the matrix plate, its study subjects, specimens, and tubes
}
\description{
Find or else create a matrix plate with plate_uid.
}
\details{
Specimen entries are named lists structured as:\preformatted{specimen_entry -> \{
  'uid': Unique study subject ID,
  'short_code': Unique study short code,
  'specimen_type': String descriptor of specimen, must be registered,
  'collection_date': Optional date of collection,
  'barcode': Matrix tube barcode,
  'well_position': Well position in plate,
  'comments': Optional comments about matrix tube
\}
}

specimen entries should be created from a CSV
}
