% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{add_study_subject}
\alias{add_study_subject}
\title{Add study subject}
\usage{
add_study_subject(uid, study_id)
}
\arguments{
\item{uid}{Unique identifier of study subject to be added to study}

\item{study_id}{integer ID of study to add study subject to}
}
\value{
returns added study subject
}
\description{
Add study subject
}
