% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{add_study_subjects}
\alias{add_study_subjects}
\title{Add study subjects}
\usage{
add_study_subjects(uids, study_id)
}
\arguments{
\item{uids}{vector of unique identifiers of study subjects to be added to
study}

\item{study_id}{integer ID of study to add study subjects to}
}
\value{
returns list of added study subjects
}
\description{
Add study subjects
}
