% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{create_location}
\alias{create_location}
\title{Create Location}
\usage{
create_location(description)
}
\arguments{
\item{description}{short unique string that describes the location}
}
\value{
Location entry
}
\description{
Register a new storage location
}
