% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{create_specimen_type}
\alias{create_specimen_type}
\title{Create Specimen Type}
\usage{
create_specimen_type(label)
}
\arguments{
\item{label}{unique description of the specimen type}
}
\value{
returns the created specimen type
}
\description{
Create Specimen Type
}
