% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{delete_location}
\alias{delete_location}
\title{Delete Location}
\usage{
delete_location(location_id)
}
\arguments{
\item{location_id}{integer ID of the location to delete}
}
\value{
Returns TRUE if the location is succesfully deleted. Raises a
ValueError if there are storage containers associated with the location.
}
\description{
Cannot delete locations with storage containers associated.
}
