% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{delete_matrix_tubes}
\alias{delete_matrix_tubes}
\title{Delete Matrix Tubes}
\usage{
delete_matrix_tubes(matrix_tubes)
}
\arguments{
\item{matrix_tubes}{List of matrix tubes to be deleted}
}
\value{

}
\description{
Delete matrix tubes and any associated specimens that now do not
have any storage containers associated
}
