% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{delete_matrix_tubes_and_specimens}
\alias{delete_matrix_tubes_and_specimens}
\title{Delete Matrix Tubes and Specimens}
\usage{
delete_matrix_tubes_and_specimens(matrix_tubes, specimens)
}
\arguments{
\item{matrix_tubes}{List of matrix tubes to be deleted}

\item{specimens}{List of specimens to be deleted}
}
\value{
returns the IDs of deleted matrix tubes and specimens
}
\description{
Simultaneously delete matrix tubes and specimens. This is
necessary to delete specimens, otherwise there will be logical errors.
}
