% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{delete_plate}
\alias{delete_plate}
\title{Delete Plate}
\usage{
delete_plate(plate_id)
}
\arguments{
\item{plate_id}{integer ID of the plate to be deleted}
}
\value{
returns TRUE if the plate is succesfully deleted. Raises ValueError
if the plate contains tubes
}
\description{
Delete Plate
}
