% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{delete_specimen_type}
\alias{delete_specimen_type}
\title{Delete Specimen Type}
\usage{
delete_specimen_type(specimen_type_id)
}
\arguments{
\item{specimen_type_id}{integer ID of the specimen type to delete}
}
\value{
returns TRUE if the type is successfully deleted. Raises ValueError
if the type has specimens associated.
}
\description{
Delete Specimen Type
}
