% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{delete_study_subject}
\alias{delete_study_subject}
\title{Delete study subject}
\usage{
delete_study_subject(study_subject_id)
}
\arguments{
\item{study_subject_id}{integer ID of study subject to delete}
}
\value{
returns TRUE if study subject is deleted. Raises NoResultFound if
study_subject_id does not match a study subject. Raises ValueError if there
are specimens associated with the sample.
}
\description{
Cannot delete study subjects that have specimens associated in
the database.
}
