% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{find_all_specimens}
\alias{find_all_specimens}
\title{Find All Specimens}
\usage{
find_all_specimens(
  specimen_entries,
  date_format = getOption("sampledblib.dateformat", "\%d/\%m/\%Y"),
  as_data_frame = TRUE
)
}
\arguments{
\item{specimen_entries}{list of specimen entries to search for}

\item{date_format}{optional date formatting string for output}

\item{as_data_frame}{logical indicating whether to return as data.frame.
Default is TRUE.}
}
\value{
list or data.frame of specimens and their locations
}
\description{
Find all specimens associated with the provided specimen
entries. Useful for exporting.
}
\details{
Specimen entries are named lists structured as:\preformatted{specimen_entry -> \{
  'uid': Unique study subject ID,
  'short_code': Unique study short code,
  'specimen_type': String descriptor of specimen, must be registered,
  'collection_date': Optional date of collection,
\}
}

specimen entries should be created from a CSV
}
