% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{find_specimens}
\alias{find_specimens}
\title{Find Specimens}
\usage{
find_specimens(
  specimen_entries,
  date_format = getOption("sampledblib.dateformat", "\%d/\%m/\%Y")
)
}
\arguments{
\item{specimen_entries}{list of specimen entries to search for}

\item{date_format}{optional date formatting string for output}
}
\value{
list of specimens and their locations
}
\description{
Find Specimens
}
\details{
Specimen entries are named lists structured as:\preformatted{specimen_entry -> \{
  'uid': Unique study subject ID,
  'short_code': Unique study short code,
  'specimen_type': String descriptor of specimen, must be registered,
  'collection_date': Optional date of collection,
\}
}

specimen entries should be created from a CSV
}
