% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_location}
\alias{get_location}
\title{Get Location}
\usage{
get_location(location_id)
}
\arguments{
\item{location_id}{integer ID of location to get}
}
\value{
Location entry
}
\description{
Get Location
}
