% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_matrix_plate}
\alias{get_matrix_plate}
\title{Get Matrix Plate}
\usage{
get_matrix_plate(plate_id)
}
\arguments{
\item{plate_id}{integer ID of the matrix plate}
}
\value{
returns the matrix plate, its study subjects, specimens, and tubes
}
\description{
Get a matrix plate along with its associated study subjects,
specimens, and tubes
}
