% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_matrix_tube}
\alias{get_matrix_tube}
\title{Get Matrix Tube}
\usage{
get_matrix_tube(barcode)
}
\arguments{
\item{barcode}{matrix tube barcode}
}
\value{
returns the matrix tube with barcode
}
\description{
Get Matrix Tube
}
