% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_matrix_tubes}
\alias{get_matrix_tubes}
\title{Get Matrix Tubes}
\usage{
get_matrix_tubes(barcodes)
}
\arguments{
\item{barcodes}{vector of matrix tube barcodes}
}
\value{
returns list of matrix tubes with provided barcodes
}
\description{
Get Matrix Tubes
}
