% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_matrix_tubes_from_specimens}
\alias{get_matrix_tubes_from_specimens}
\title{Get Matrix Tubes from Specimens}
\usage{
get_matrix_tubes_from_specimens(specimen_entries)
}
\arguments{
\item{specimen_entries}{list of specimen entries to search for}
}
\value{
returns list of matrix tubes for specimens
}
\description{
Get Matrix Tubes from Specimens
}
\details{
Specimen entries are named lists structured as:\preformatted{specimen_entry -> \{
  'uid': Unique study subject ID,
  'short_code': Unique study short code,
  'specimen_type': String descriptor of specimen, must be registered,
  'collection_date': Optional date of collection,
\}
}

specimen entries should be created from a CSV
}
