% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_specimen_type}
\alias{get_specimen_type}
\title{Get Specimen Type}
\usage{
get_specimen_type(specimen_type_id)
}
\arguments{
\item{specimen_type_id}{integer ID of the specimen type}
}
\value{
returns the specimen type with ID
}
\description{
Get Specimen Type
}
