% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_specimens}
\alias{get_specimens}
\title{Get Specimens}
\usage{
get_specimens(uid, short_code, collection_date = NULL)
}
\arguments{
\item{uid}{unique identifier for a sample within a study}

\item{short_code}{unique short code identifying the study}

\item{collection_date}{optional collection date}
}
\value{
list of Specimens
}
\description{
Get specimens associated with a sample
}
