% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_study}
\alias{get_study}
\title{Get study}
\usage{
get_study(id)
}
\arguments{
\item{id}{id of the study}
}
\value{
returns list containing the study, its subjects, specimens, and their
containers
}
\description{
Get study
}
