% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_study_by_short_code}
\alias{get_study_by_short_code}
\title{Get study by short code}
\usage{
get_study_by_short_code(short_code)
}
\arguments{
\item{short_code}{unique short code that is used by the study}
}
\value{
returns the study with the short code
}
\description{
Get study by short code
}
