% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{get_study_subjects}
\alias{get_study_subjects}
\title{Get Study Subjects}
\usage{
get_study_subjects(id)
}
\arguments{
\item{id}{integer ID of study to search}
}
\value{
returns list of study subjects in study
}
\description{
Get study subjects associated with provided study ID
}
