% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{initialize_sampledb}
\alias{initialize_sampledb}
\title{Initialize SampleDB}
\usage{
initialize_sampledb(
  conn_string = getOption("sampledblib.db.location", ""),
  date_format = getOption("sampledblib.fm.dateformat", "\%d-\%b-\%Y")
)
}
\value{

}
\description{
Initialize SampleDB
}
