% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_barcode_file}
\alias{parse_barcode_file}
\title{Parse Barcode File}
\usage{
parse_barcode_file(filepath, expand_path = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
Vector of barcodes in file
}
\description{
Parse Barcode File
}
\details{
Extracts the barcode column from a csv file
}
