% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_barcode_search_file}
\alias{parse_barcode_search_file}
\title{Parse Barcode Search File}
\usage{
parse_barcode_search_file(filepath, expand_path = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
a list containing a vector of barcode entries, the original fields of the data, and the index of the barcode column
}
\description{
Parses a file that satisfies the Barcode Search format
}
\details{
A csv file satisfies the Barcode Search format if it contains a column called "Barcode". This is useful for converting data where the barcode is known into fully expanded sample information.
}
