% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_new_plate_file}
\alias{parse_new_plate_file}
\title{Parse New Plate File}
\usage{
parse_new_plate_file(filepath, expand_path = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
returns list of specimen entries with locations from file
}
\description{
Parses a file that satisfies the New Plate format
}
\details{
A csv file is New Plate format if it contains the columns:\tabular{ccccccc}{
   Well \tab Barcode \tab UID \tab Specimen Type \tab Study Short Code \tab Date (optional) \tab Comments (optional) \cr
   ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
}
}
