% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_plate_update_files}
\alias{parse_plate_update_files}
\title{Parse Plate Update Files}
\usage{
parse_plate_update_files(filepaths, expand_path = FALSE)
}
\arguments{
\item{filepaths}{vector of file paths pointing to plate files to be processed}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
returns list of matrix tube entries
}
\description{
Parses files that are of the Plate Update format
}
\details{
Matrix tube entries are named lists structured as:\preformatted{  matrix_tube_entry -> \{
  'barcode': Unique tube barcode,
  'plate_uid': Unique Plate ID matrix tube is going into,
  'well_position': New matrix tube position,
  'comments': Optional comments
  \}
}

This is useful for taking CSVs to update tube locations
}
