% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_specimen_search_file}
\alias{parse_specimen_search_file}
\title{Parse Specimen Search File}
\usage{
parse_specimen_search_file(filepath, expand_path = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
returns list of specimen entries
}
\description{
Parses a file that satisfies the Specimen Search format
}
\details{
A csv file is Specimen Search format if it contains the columns:
| UID | Specimen Type | Study Short Code | Date (optional) |
| :---: | :---: | :---: | :---: |
| ... | ... | ... | ... |
}
