% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_study_subject_file}
\alias{parse_study_subject_file}
\title{Parse Study Subject File}
\usage{
parse_study_subject_file(filepath, expand_path = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
returns vector of study subject UIDs
}
\description{
extracts the UID field from a Study Subject file
}
