% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{parse_subject_search_file}
\alias{parse_subject_search_file}
\title{Parse Subject Search File}
\usage{
parse_subject_search_file(filepath, expand_path = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{expand_path}{logical whether or not to expand a path from '~' to the
fully qualified name}
}
\value{
returns list of subject entries
}
\description{
Parses a file that satisfies the Subject Search format
}
\details{
A csv file is Specimen Search format if it contains the columns:
| UID | Study Short Code | Date (optional) |
| :---: | :---: | :---: | :---: |
| ... | ... | ... | ... |
}
