% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manager.R
\name{set_date_format}
\alias{set_date_format}
\title{Set Date Format}
\usage{
set_date_format(date_format)
}
\arguments{
\item{date_format}{}
}
\description{
Set the date format string of the file parser
}
