% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{set_matrix_tubes_exhausted}
\alias{set_matrix_tubes_exhausted}
\title{Set Matrix Tubes Exhausted}
\usage{
set_matrix_tubes_exhausted(barcodes)
}
\arguments{
\item{barcodes}{vector of matrix tube barcodes to set exhausted}
}
\value{
returns list of matrix tubes that have been set exhausted
}
\description{
Set Matrix Tubes Exhausted
}
