% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{unset_matrix_tubes_exhausted}
\alias{unset_matrix_tubes_exhausted}
\title{Unset Matrix Tubes Exhausted}
\usage{
unset_matrix_tubes_exhausted(barcodes)
}
\arguments{
\item{barcodes}{vector of matrix tube barcodes to unset exhausted}
}
\value{
returns list of matrix tubes that have been unset exhausted
}
\description{
Unset Matrix Tubes Exhausted
}
