% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{update_location}
\alias{update_location}
\title{Update Location}
\usage{
update_location(location_id, d)
}
\arguments{
\item{location_id}{integer ID of the location to update}

\item{d}{named list containing fields to update}
}
\value{
Updated location entry
}
\description{
Update Location
}
