% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{update_matrix_tube_locations}
\alias{update_matrix_tube_locations}
\title{Update matrix tube locations}
\usage{
update_matrix_tube_locations(matrix_tube_entries)
}
\arguments{
\item{matrix_tube_entries}{list of matrix tube entries}
}
\value{
returns named list containing updated matrix plates, their study
subjects, specimens, and tubes
}
\description{
Update matrix tube locations
}
\details{
Matrix tube entries are named lists structured as:\preformatted{matrix_tube_entry -> \{
 'barcode': Unique tube barcode,
 'plate_uid': Unique Plate ID matrix tube is going into,
 'well_position': New matrix tube position,
 'comments': Optional comments \}
}

matrix tube entries should be created from a CSV
}
