% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{update_specimen_type}
\alias{update_specimen_type}
\title{Update Specimen Type}
\usage{
update_specimen_type(specimen_type_id, d)
}
\arguments{
\item{specimen_type_id}{integer ID of the specimen type to update}

\item{d}{named list of fields to update}
}
\value{
returns the updated specimen type
}
\description{
Update Specimen Type
}
