% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledb.R
\name{update_study}
\alias{update_study}
\title{Update Study}
\usage{
update_study(study_id, d)
}
\arguments{
\item{study_id}{integer ID of study to update}

\item{d}{named list containing fields to update}
}
\value{
returns the updated study, its subjects, specimens, and their
containers
}
\description{
Update Study
}
\examples{
update_study(1, list(title = "My new title"))
}
